rename AppBasicMac -> AppMac, AppBasicMSw -> AppMsw
